requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
python-json-logger>=2.0.7
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
chromadb>=0.5.11
sentence-transformers>=2.7.0
numpy>=1.26.0
pypdf>=4.0.0
pymupdf>=1.24.10
openai>=1.48.0
tiktoken>=0.7.0
syntok>=1.4.4
ragas>=0.1.9
datasets>=2.19.0
